\name{updateData}
\alias{updateData}
\title{Update data from the U.S. CDC}
\usage{
	updateData(
		Data,
		what="all")
}
\arguments{
	\item{Data}{data from \code{\link{Data}}.}
	\item{what}{The thing to update. Defaults to \code{"all"}, otherwise can be the names of one or more of the elements of \code{Data}.}
}
\description{
	Updates data from the web.
}
\value{
	An updated data object.
}
\references{
	References here.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
